requirements.txt
# Runtime dependencies for PEIRASTIC_control.
pyzmq>=18.0
numpy>=1.18
PyYAML>=5.0
easydict>=1.9
termcolor>=1.1
scipy>=1.5
protobuf>=3.20.0,<3.21.0
hidapi>=0.14
glfw
matplotlib
rospkg
netifaces>=0.11.0
